clear media
