+++ ios/Sources/AppFeature/AppFeature.swift
@@ -23,22 +23,29 @@ public enum AppAction {
     case appTab(AppTabAction)
 }
 
-public let appReducer = Reducer<AppState, AppAction, AppEnvironment> { state, action, environment in
-    switch action {
-    case .refresh:
-        return environment.feedRepository.feedContents()
-            .catchToEffect()
-            .map(AppAction.refreshResponse)
-    case .needRefresh:
-        state = .needToInitialize
-        return .none
-    case let .refreshResponse(.success(feedContents)):
-        state = .initialized(.init(feedContents: feedContents))
-        return .none
-    case let .refreshResponse(.failure(error)):
-        state = .errorOccurred
-        return .none
-    case .appTab:
-        return .none
+public let appReducer = Reducer<AppState, AppAction, AppEnvironment>.combine(
+    appTabReducer.pullback(
+        state: /AppState.initialized,
+        action: /AppAction.appTab,
+        environment: { $0 }
+    ),
+    .init { state, action, environment in
+        switch action {
+        case .refresh:
+            return environment.feedRepository.feedContents()
+                .catchToEffect()
+                .map(AppAction.refreshResponse)
+        case .needRefresh:
+            state = .needToInitialize
+            return .none
+        case let .refreshResponse(.success(feedContents)):
+            state = .initialized(.init(feedContents: feedContents))
+            return .none
+        case let .refreshResponse(.failure(error)):
+            state = .errorOccurred
+            return .none
+        case .appTab:
+            return .none
+        }
     }
-}
+)
