+++ ios/Sources/MediaFeature/MediaType.swift
@@ -1,0 +1,5 @@
+public enum MediaType {
+    case blog
+    case video
+    case podcast
+}
